+++ lib/Images/Jpeg.cpp
@@ -3,7 +3,7 @@
 #include <cstdint>
 #include <stdexcept>
 
-#include "Utils/Handlers.hpp"
+#include "Utils/Files/Handlers.hpp"
 
 
 namespace Images {
