+++ tests/CMakeLists.txt
@@ -1,0 +1,9 @@
+include(DoctestHelpers)
+
+add_doctest(
+    NAME test_libLoadersObj
+    SOURCES
+    lib/Loaders/test_libLoadersObj.cpp
+    LIBS
+    Loaders::Obj
+)
